This commit allows Aleph to be used as the ring handler.
It breaks the configuration options for the serve task
available in previous releases. Instead of a boolean option
for :httpkit, server now takes a :server key for which valid
values are :httpkit or :alpeh.

There are no default ring wrappers for Aleph handlers.

For now the caller will have to wrap their handler with
required middleware. This situation could be improved.

uses -S --server switch to indicate server to use on command line.
Default is Jetty